+++ pkg/envoy/registry/announcement_handlers.go
@@ -35,7 +35,7 @@ func (pr *ProxyRegistry) ReleaseCertificateHandler(certManager certificateReleas
 			}
 
 			proxyUUID := deletedPodObj.Labels[constants.EnvoyUniqueIDLabelName]
-			if proxy, ok := pr.connectedProxies[proxyUUID]; ok {
+			if proxy := pr.GetConnectedProxy(proxyUUID); proxy != nil {
 				log.Warn().Msgf("Pod with label %s: %s found in proxy registry; releasing certificate for proxy %s", constants.EnvoyUniqueIDLabelName, proxyUUID, proxy.Identity)
 				certManager.ReleaseCertificate(envoy.NewXDSCertCNPrefix(proxy.UUID, proxy.Kind(), proxy.Identity))
 			} else {
